[major] more modules: date-utils, intervals
